refactor: split demo components and optimize posts
