+++ web/src/components/examples/index.ts
@@ -1,0 +1,6 @@
+export { default as BasicPostForm } from './BasicPostForm';
+export { default as OptimisticPostList } from './OptimisticPostList';
+export { default as NewsletterForm } from './NewsletterForm';
+export { default as SearchPosts } from './SearchPosts';
+export { default as FileUploadForm } from './FileUploadForm';
+export { default as Dashboard } from './Dashboard';
